Refractor MedianOfTwoSortedArrays solution and test, remove .gradle folder
