+++ .gitignore
@@ -24,3 +24,4 @@ hs_err_pid*
 
 build
 .idea
+.gradle
