--- .gradle/buildOutputCleanup/cache.properties
@@ -1,2 +1,0 @@
-#Wed Feb 09 01:53:25 PST 2022
-gradle.version=7.1
