+++ src/test/kotlin/com/albion/labs/numbers/MedianOfTwoSortedArraysTest.kt
@@ -11,7 +11,7 @@ class MedianOfTwoSortedArraysTest {
         val nums2 = intArrayOf(2)
         val expected = 2.0
         val actual = MedianOfTwoSortedArrays.findMedianSortedArrays(nums1, nums2)
-        assertEquals(actual, expected, 0.0)
+        assertEquals(actual, expected, 0.000001)
     }
 
     @Test
@@ -20,6 +20,15 @@ class MedianOfTwoSortedArraysTest {
         val nums2 = intArrayOf(2,4)
         val expected = 2.5
         val actual = MedianOfTwoSortedArrays.findMedianSortedArrays(nums1, nums2)
-        assertEquals(actual, expected, 0.0)
+        assertEquals(actual, expected, 0.000001)
+    }
+
+    @Test
+    fun testBasicV2() {
+        val nums1 = intArrayOf(1,3)
+        val nums2 = intArrayOf(2,4)
+        val expected = 2.5
+        val actual = MedianOfTwoSortedArrays.findMedianSortedArraysV2(nums1,nums2)
+        assertEquals(actual, expected, 0.000001)
     }
 }
